+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 [Diff](https://github.com/kvz/lanyon/compare/v0.0.33...v0.0.34)
 
+- [x] Allow asset loaders to load from `bower_components`
 - [x] Reset (a possibly corrupted) `records.json`
 - [x] Fix: `Module not found: Error: Cannot resolve module 'fs'` by adding  `node: { fs: 'empty' }, target: 'node'` to webpack config
 - [x] Support `prebuild*` hooks being arrays